remove test for service specific conflated middleware
